ifrc-api/requirements.txt
pycountry==18.5.26
requests==2.20.0
python-dateutil==2.7.3
numpy==1.15.0
pandas==0.23.4
grequests==0.3.0
dateparser==0.7.0
aiohttp==3.3.2
python-dotenv==0.9.1
click==6.7
cdsapi==0.1.1
#git+git://github.com/jswhit/pyproj@40bc5389e7ee8212d9f5a790f582970007e367e1
#pygrib==2.0.3
pandas==0.23.4
